+++ anvil/src/udev.rs
@@ -738,7 +738,7 @@ fn render_surface(
         return Ok(());
     };
 
-    let dmabuf = surface.surface.next_buffer()?;
+    let (dmabuf, _age) = surface.surface.next_buffer()?;
     renderer.bind(dmabuf)?;
 
     // and draw to our buffer
@@ -867,7 +867,7 @@ fn schedule_initial_render<Data: 'static>(
 }
 
 fn initial_render(surface: &mut RenderSurface, renderer: &mut Gles2Renderer) -> Result<(), SwapBuffersError> {
-    let dmabuf = surface.next_buffer()?;
+    let (dmabuf, _age) = surface.next_buffer()?;
     renderer.bind(dmabuf)?;
     // Does not matter if we render an empty frame
     renderer